+++ ChangeLog.md
@@ -4,6 +4,7 @@
 
     [ 20140806 ]
     * Improved Script reader and its Unit Test
+    * Successfully parsing the TR2 Script strings!
 
     [ 20140804 ]
     * Added preliminary Tomb Raider Script parser